ecs: Task replanning (#844)
* ecs: Preserve system entries in runner

* ecs: Store systemCount on runner

* jobs: Specify 16 byte alignment for task data

* jobs: Fix inconsistent name in graph tests

* jobs: Add jobs_graph_copy

* ecs: Use a temp job graph instead of temp sys entries

* jobs: Store task names in bump alloc

* ecs: Fix formatting

* jobs: Remove heap allocations from jobs_graph_topologically_insert

* jobs: Refactor JobsTaskId to 16 bit

* jobs: Save parentsCounts as 16 bit

* jobs: Replace some scratch allocations with stack allocations

* ecs: Introduce runner plans

* ecs: Add optional plan shuffle

* ecs: Compute runner plan cost

* ecsL Pick the best plan

* ecs: Minor cleanup in runner func names

* ecs: Minor cleanup in runner struct names

* ecs: Remove explicit storing of sys order

* ecs: Avoid storing sysId in runner

* ecs: Remove unneeded RunnerSystemEntry struct

* ecs: Refactor runner sys collection

* ecs: Runner store systems on the stack

* ecs: Remove the need to store the systemCount

* ecs: Reduce padding in EcsRunner struct

* ecs: Add runner replan task

* check,ecs: Fix MSVC compat

* ecs: Fix missing const mod

* ecs: Skip replan if single-threaded

* ecs: Add log when finalizing a new plan

* ecs: Add feature to disable replanning

* jobs: Add jobs_dot_dump_graph_to_path_default util

* debug: Add 'dump graph' button to ecs panel

+++ libs/jobs/src/executor.c
@@ -173,7 +173,7 @@ static void executor_perform_work(const JobWorkerId wId, WorkItem item) {
    */
 
   JobTaskId childTasks[job_max_task_children];
-  usize     childCount = 0;
+  u32       childCount = 0;
   jobs_graph_for_task_child(item.job->graph, item.task, child) {
     diag_assert_msg(
         childCount < job_max_task_children,
@@ -326,7 +326,7 @@ void executor_run(Job* job) {
   diag_assert_msg(g_jobsWorkerCount, "Job system has to be initialized jobs_init() first.");
 
   const JobWorkerId wId           = g_jobsWorkerId;
-  const usize       rootTaskCount = jobs_graph_task_root_count(job->graph);
+  const u32         rootTaskCount = jobs_graph_task_root_count(job->graph);
 
   /**
    * Push work items for all root tasks in the job.
@@ -336,7 +336,7 @@ void executor_run(Job* job) {
    */
 
   JobTaskId taskId = 0;
-  for (usize pushedRootTaskCount = 0; pushedRootTaskCount != rootTaskCount; ++taskId) {
+  for (u32 pushedRootTaskCount = 0; pushedRootTaskCount != rootTaskCount; ++taskId) {
     if (jobs_graph_task_has_parent(job->graph, taskId)) {
       continue; // Not a root task.
     }
